+++ src/main/java/com/trybe/gestaotime/dao/JogadorDao.java
@@ -55,7 +55,7 @@ public class JogadorDao extends GenericDao<Jogador, Integer> {
   }
 
   @Override
-  public void deletar(Long id) {
+  public void deletar(Integer id) {
     EntityManager em = GenericDao.emf.createEntityManager();
     em.getTransaction().begin();
     Jogador document = em.find(Jogador.class, id);
