+++ src/main/java/com/trybe/gestaotime/dao/TorcedorDao.java
@@ -55,7 +55,7 @@ public class TorcedorDao extends GenericDao<Torcedor, Integer> {
   }
 
   @Override
-  public void deletar(Long id) {
+  public void deletar(Integer id) {
     EntityManager em = GenericDao.emf.createEntityManager();
     em.getTransaction().begin();
     Torcedor document = em.find(Torcedor.class, id);
